refactor: move toolsets to the new architecture
