+++ front/lib/api/actions/servers/toolsets/index.ts
@@ -1,0 +1,25 @@
+import type { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
+
+import { makeInternalMCPServer } from "@app/lib/actions/mcp_internal_actions/utils";
+import { registerTool } from "@app/lib/actions/mcp_internal_actions/wrappers";
+import type { AgentLoopContextType } from "@app/lib/actions/types";
+import { TOOLSETS_TOOL_NAME } from "@app/lib/api/actions/servers/toolsets/metadata";
+import { TOOLS } from "@app/lib/api/actions/servers/toolsets/tools";
+import type { Authenticator } from "@app/lib/auth";
+
+function createServer(
+  auth: Authenticator,
+  agentLoopContext?: AgentLoopContextType
+): McpServer {
+  const server = makeInternalMCPServer("toolsets");
+
+  for (const tool of TOOLS) {
+    registerTool(auth, agentLoopContext, server, tool, {
+      monitoringName: TOOLSETS_TOOL_NAME,
+    });
+  }
+
+  return server;
+}
+
+export default createServer;
